Calculate, Save and Get achieved points of answered qes from recruiter

+++ backend/src/main/java/de/kit/recruitergame/Repository/RecruiterAnswerRepository.java
@@ -3,5 +3,5 @@ package de.kit.recruitergame.Repository;
 import de.kit.recruitergame.model.RecruiterAnswer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecruiterAnswerRepository extends JpaRepository<RecruiterAnswer,Integer> {
+public interface RecruiterAnswerRepository extends JpaRepository<RecruiterAnswer,Long> {
 }
